+++ packages/adminui-pub/CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
   - en i18n: "Select from dictionary" -> "Select from the dictionary"
+  - for `UBBaseComboBox` with applied filters (new Ext.util.Filter) applied "Select from the dictionary" action will
+    filter Vue based dictionary in the same way as in the ComboBox store (the same behavior as with obsolete Ext-based dictionary)
 
 ## [5.13.15] - 2020-05-27
 ### Added